added seperate files for the i2c dummies
