+++ test/unit/CMakeLists.txt
@@ -23,7 +23,8 @@ target_link_libraries(unit-test_network unity)
 ## ADD EXECUTABLE: unit-test_sht3x
 add_executable(unit-test_sht3x
         ${CMAKE_CURRENT_LIST_DIR}/test_sht3x.c
-        ${CMAKE_SOURCE_DIR}/src/sht3x/sht3x.c)
+        ${CMAKE_SOURCE_DIR}/src/sht3x/sht3x.c
+        ${CMAKE_CURRENT_LIST_DIR}/dummies/i2c/i2c.c)
 target_include_directories(unit-test_sht3x PUBLIC
         ${CMAKE_CURRENT_LIST_DIR}/dummies
         ${CMAKE_SOURCE_DIR}/src
