+++ test/unit/test_sht3x.c
@@ -3,113 +3,11 @@
 //
 
 #include "i2c/i2c.h"
+#include "i2c/i2c_test.h"
 #include "sht3x/sht3x_public.h"
 #include "unity.h"
 #include <stdlib.h>
 
-
-/* region VARIABLES */
-
-static const uint8_t byteZero            = 0xBE;
-static const uint8_t byteOne             = 0xEF;
-static const uint8_t correctByteChecksum = 0x92;
-static const uint8_t wrongByteChecksum   = 0x00;
-
-/* endregion */
-/* region DUMMIES FOR I2C */
-
-void I2C_Init ( i2c_inst_t * i2cHost, uint32_t baudrate, uint8_t sdaGPIO, uint8_t sclGPIO )
-  {
-    /* Nothing to do here !!
-     * Just here to satisfy compiler
-     */
-  }
-
-static I2C_ErrorCode (* I2C_WriteCommand_ptr) ( const uint8_t * commandBuffer, uint16_t sizeOfCommandBuffer, uint8_t slaveAddress, i2c_inst_t * i2cHost );
-
-I2C_ErrorCode I2C_WriteCommand ( const uint8_t * commandBuffer, uint16_t sizeOfCommandBuffer, uint8_t slaveAddress, i2c_inst_t * i2cHost )
-  {
-    return I2C_WriteCommand_ptr ( commandBuffer, sizeOfCommandBuffer, slaveAddress, i2cHost );
-  }
-
-static I2C_ErrorCode I2C_WriteCommand_Hardware_defect ( const uint8_t * commandBuffer, uint16_t sizeOfCommandBuffer, uint8_t slaveAddress, i2c_inst_t * i2cHost
-                                                      )
-  {
-    return I2C_INIT_ERROR;
-  }
-
-static I2C_ErrorCode I2C_WriteCommand_ACK_missing ( const uint8_t * commandBuffer, uint16_t sizeOfCommandBuffer, uint8_t slaveAddress, i2c_inst_t * i2cHost )
-  {
-    return I2C_ACK_ERROR;
-  }
-
-static I2C_ErrorCode I2C_WriteCommand_Pass ( const uint8_t * commandBuffer, uint16_t sizeOfCommandBuffer, uint8_t slaveAddress, i2c_inst_t * i2cHost )
-  {
-    return I2C_NO_ERROR;
-  }
-
-static I2C_ErrorCode (* I2C_ReadCommand_ptr) ( uint8_t * readBuffer, uint8_t sizeOfReadBuffer, uint8_t slaveAddress, i2c_inst_t * i2cHos );
-
-I2C_ErrorCode I2C_ReadData ( uint8_t * readBuffer, uint8_t sizeOfReadBuffer, uint8_t slaveAddress, i2c_inst_t * i2cHost )
-  {
-    return I2C_ReadCommand_ptr ( readBuffer, sizeOfReadBuffer, slaveAddress, i2cHost );
-  }
-
-static I2C_ErrorCode I2C_ReadCommand_Hardware_defect ( uint8_t * readBuffer, uint8_t sizeOfReadBuffer, uint8_t slaveAddress, i2c_inst_t * i2cHost )
-  {
-    return I2C_INIT_ERROR;
-  }
-
-static I2C_ErrorCode I2C_ReadCommand_ACK_missing ( uint8_t * readBuffer, uint8_t sizeOfReadBuffer, uint8_t slaveAddress, i2c_inst_t * i2cHost )
-  {
-    return I2C_ACK_ERROR;
-  }
-
-static I2C_ErrorCode I2C_ReadCommand_Pass ( uint8_t * readBuffer, uint8_t sizeOfReadBuffer, uint8_t slaveAddress, i2c_inst_t * i2cHost )
-  {
-    /**
-     * @brief failing I²C implementation
-     *
-     * I2C only writes one byte,
-     * this implementation breaks this rule by always writing three bytes in a row
-     *
-     * FIXME: Improve implementation to match I2C specifications
-     */
-    
-    /* generate sample data without any real world connection to test implementation */
-    for ( uint8_t index = 0 ; index < sizeOfReadBuffer ; index = index + 3 )
-      {
-        readBuffer[ index ]     = byteZero;
-        readBuffer[ index + 1 ] = byteOne;
-        readBuffer[ index + 2 ] = correctByteChecksum;
-      }
-    
-    return 0x00;
-  }
-
-static I2C_ErrorCode I2C_ReadCommand_provoke_checksum_fail ( uint8_t * readBuffer, uint8_t sizeOfReadBuffer, uint8_t slaveAddress, i2c_inst_t * i2cHost )
-  {
-    /**
-     * @brief failing I2C implementation
-     *
-     * I2C only writes one byte,
-     * this implementation breaks this rule by always writing three bytes in a row
-     *
-     * FIXME: Improve implementation to match I2C specifications
-     */
-    
-    /* generate sample data without any real world connection to test implementation */
-    for ( uint8_t index = 0 ; index < sizeOfReadBuffer ; index = index + 3 )
-      {
-        readBuffer[ index ]     = byteZero;
-        readBuffer[ index + 1 ] = byteOne;
-        readBuffer[ index + 2 ] = wrongByteChecksum;
-      }
-    
-    return 0x00;
-  }
-
-/* endregion */
 
 void setUp ( void )
   {
